requirements.txt
streamlit==1.45.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
matplotlib==3.10.3
seaborn==0.13.2
xgboost==3.0.0
shap==0.47.2
joblib==1.5.0
Pillow==11.2.1
slicer==0.0.8